Pull out vector only logic into new component
