+++ web/src/features/map/Map.tsx
@@ -3,7 +3,7 @@ import 'ol/ol.css'
 import React, { useRef, useState, useEffect, useCallback } from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 import * as ol from 'ol'
-import { get, toLonLat } from 'ol/proj'
+import { toLonLat } from 'ol/proj'
 import { FeatureLike } from 'ol/Feature'
 import OLOverlay from 'ol/Overlay'
 import { MapOptions } from 'ol/PluggableMap'
@@ -11,52 +11,14 @@ import { defaults as defaultControls } from 'ol/control'
 
 import { Button, ErrorBoundary } from 'components'
 import { ObjectEvent } from 'ol/Object'
-import VectorLayer from 'features/map/VectorLayer'
-import GeoJSON from 'ol/format/GeoJSON'
-import * as olSource from 'ol/source'
-import { selectFireWeatherStations } from 'app/rootReducer'
-import { useDispatch, useSelector } from 'react-redux'
+import { useDispatch } from 'react-redux'
 import { AccuracyWeatherVariableEnum } from 'features/fireWeather/components/AccuracyVariablePicker'
-import { fetchWxStations, selectStation } from 'features/stations/slices/stationsSlice'
-import {
-  computeRHAccuracyColor,
-  computeRHAccuracySize,
-  computeStroke,
-  computeTempAccuracyColor,
-  computeTempAccuracySize
-} from 'features/fireWeather/components/maps/stationAccuracy'
-import { Style, Fill } from 'ol/style'
-import CircleStyle from 'ol/style/Circle'
-import { getDetailedStations, StationSource } from 'api/stationAPI'
+import { selectStation } from 'features/stations/slices/stationsSlice'
+import FireIndicesVectorLayer from 'features/fireWeather/components/maps/FireIndicesVectorLayer'
 
 const zoom = 6
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const rhPointStyleFunction = (feature: any) => {
-  const rhPointColor = computeRHAccuracyColor(feature.values_)
-  return new Style({
-    image: new CircleStyle({
-      radius: computeRHAccuracySize(feature.values_),
-      fill: new Fill({ color: rhPointColor }),
-      stroke: computeStroke(rhPointColor)
-    })
-  })
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const tempPointStyleFunction = (feature: any) => {
-  const tempPointColor = computeTempAccuracyColor(feature.values_)
-  return new Style({
-    image: new CircleStyle({
-      radius: computeTempAccuracySize(feature.values_),
-      fill: new Fill({ color: tempPointColor }),
-      stroke: computeStroke(tempPointColor)
-    })
-  })
-}
-
 export const MapContext = React.createContext<ol.Map | null>(null)
-
 export interface RedrawCommand {
   redraw: boolean
 }
@@ -116,19 +78,6 @@ const Map = ({
   const [map, setMap] = useState<ol.Map | null>(null)
   const [feature, setFeature] = useState<FeatureLike | null>(null)
   const [currentCenter, setCurrentCenter] = useState(center)
-
-  const { stations } = useSelector(selectFireWeatherStations)
-
-  useEffect(() => {
-    dispatch(
-      fetchWxStations(getDetailedStations, StationSource.unspecified, toiFromQuery)
-    )
-  }, []) // eslint-disable-line react-hooks/exhaustive-deps
-
-  const styleFunction =
-    selectedWxVariable === AccuracyWeatherVariableEnum['Relative Humidity']
-      ? rhPointStyleFunction
-      : tempPointStyleFunction
 
   // on component mount
   useEffect(() => {
@@ -254,19 +203,9 @@ const Map = ({
       <MapContext.Provider value={map}>
         <div ref={mapRef} className={classes.map} data-testid="map">
           {children}
-          <VectorLayer
-            source={
-              new olSource.Vector({
-                features: new GeoJSON().readFeatures(
-                  { type: 'FeatureCollection', features: stations },
-                  {
-                    featureProjection: get('EPSG:3857')
-                  }
-                )
-              })
-            }
-            style={styleFunction}
-            zIndex={1}
+          <FireIndicesVectorLayer
+            toiFromQuery={toiFromQuery}
+            selectedWxVariable={selectedWxVariable}
           />
         </div>
         {renderTooltip && (
